Add ORT_HOME_PATH/cache to the locations for the GH cache
Some package managers stores downloaded artifacts in the ORT_HOME_PATH/cache directory for faster access after the first run, one example being the maven package manager. 

Adding the .m2 directory as cache is not sufficient in this case, as the package manager will do a head request for all artifacts all the time unless it is found in the aforementioned cache.

Signed-off-by: Thomas Neidhart <thomas.neidhart@gmail.com>